forecasting-example/components/model_evaluation/requirements.txt
sklearn
pandas
fire
pickle
google-cloud-storage
